requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
blis==0.7.9
boto3==1.33.13
botocore==1.33.13
Brotli==1.1.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorgram.py==1.2.0
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-filter==23.5
django-js-asset==2.0.0
django-modelcluster==6.1
django-permissionedforms==0.1
django-quill-editor==0.1.40
django-social-share==2.3.0
django-storages==1.14.2
django-taggit==4.0.0
django-tinymce==3.6.1
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
fonttools==4.39.3
gunicorn==21.2.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
l18n==2021.3
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
murmurhash==1.0.9
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==2.1.3
pathy==0.10.1
Pillow==9.4.0
pillow-heif==0.14.0
preshed==3.0.8
prettytable==3.7.0
psycopg2-binary==2.9.9
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
s3transfer==0.8.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
telepath==0.3.1
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wagtail==5.2.2
wasabi==1.1.1
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.6.3
